refactor: split into modules, add enum declaration support
Splits the signal, message and overall codegen logic into their
own modules.  Adds tests for `CYCLE_TIME` generated constants,
enum-based declarations, and incomplete DBC files.

Adds the ability to declare types using an `enum` in addition to
using `struct`.  This can be more convenient for cases where the
generated types are used standalone.
