+++ tests/incomplete.dbc
@@ -1,0 +1,7 @@
+VERSION "1"
+
+NS_ :
+
+BS_:
+
+BA_DEF_DEF_REL_ "Incomplete" "";
